requirements.txt
pip~=22.0.2
toml~=0.10.2
wheel~=0.37.1
google~=3.0.0
cryptography~=3.4.8
lxml~=4.8.0
keyring~=23.5.0
setuptools~=60.2.0
future~=0.18.2
docutils~=0.17.1
Pillow~=9.0.1